DOC: Use `npx` instead of `onp` and fix initial readme note a bit
`onp` was just me thinking I remembered us busing it in the dev
meeting a long time ago. If it was used in this context, nobody
ever had any meaning for it, and indeed `onp` is sometimes used
for "original" numpy.
So, `npx` is likely not the final word for a "preferred" spelling,
but it is much better than `onp` :).